# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Askito Rest api",
      "description": "A community for developers to share their knowledge"
    },
    "host": "askito.herokuapp.com",
    "basePath": "/api/",
...
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/posts/admin/edit/title/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update post title by id",
          "description": "",
          "operationId": "updatePostTitleById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Post title that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
      "/posts/admin/update/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update post",
          "description": "",
          "operationId": "updatePostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Post"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/admin/edit/data/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update post data by id",
          "description": "",
          "operationId": "updatePostDataById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "data",
              "description": "Post data that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/admin/edit/tags/{id}": {
        "put": {
          "tags": ["Admin"],
          "summary": "update post tags by id",
          "description": "",
          "operationId": "updatePostTagsById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "tags",
              "type": "array",
              "description": "Post tags that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
      "/posts/add": {
        "post": {
          "tags": ["Posts"],
          "summary": "add new post",
          "description": "",
          "operationId": "addPost",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Post"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "post created successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/{id}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get posts by id",
          "description": "",
          "operationId": "getPostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get total posts",
          "description": "",
          "operationId": "getPosts",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "No posts found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/myposts/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get total posts of the login user",
          "description": "",
          "operationId": "getPostsOfLoginUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "No posts found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/update/{id}": {
        "put": {
          "tags": ["Posts"],
          "summary": "update post",
          "description": "",
          "operationId": "updatePostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Post"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/edit/title/{id}": {
        "put": {
          "tags": ["Posts"],
          "summary": "update post title by id",
          "description": "",
          "operationId": "updatePostTitleById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "title",
              "description": "Post title that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/edit/data/{id}": {
        "put": {
          "tags": ["Posts"],
          "summary": "update post data by id",
          "description": "",
          "operationId": "updatePostDataById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "data",
              "description": "Post data that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/edit/tags/{id}": {
        "put": {
          "tags": ["Posts"],
          "summary": "update post tags by id",
          "description": "",
          "operationId": "updatePostTagsById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "tags",
              "description": "Post tags that needs to be updated",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/delete/{id}": {
        "delete": {
          "tags": ["Posts"],
          "summary": "delete post by id",
          "description": "",
          "operationId": "deletePostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "post deleted"
            },
            "401": {
              "description": "access denied."
            },
            "404": {
              "description": "The post with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Post": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "data": {
              "type": "string"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
            
          }
        },
...
      } 
}